Added numerical method for CF

+++ distributions/univariate/continuous/NormalRand.cpp
@@ -39,7 +39,7 @@ bool NormalRand::setupTables()
     /// implicit value for the top layer
     stairWidth[256] = 0;
 
-    for (unsigned i = 1; i <= 255; ++i)
+    for (size_t i = 1; i <= 255; ++i)
     {
         /// such y_i that f(x_{i+1}) = y_i
         stairWidth[i] = std::sqrt(-2 * std::log(stairHeight[i - 1]));
