+++ distributions/univariate/discrete/BetaBinomialRand.h
@@ -29,7 +29,7 @@ public:
     double Mean() const override;
     double Variance() const override;
 
-    double Mode() const override;
+    int Mode() const override;
     double Skewness() const override;
     double ExcessKurtosis() const override;
 };
