+++ distributions/univariate/discrete/UniformDiscreteRand.h
@@ -32,7 +32,7 @@ public:
     double Variance() const override;
 
     double Median() const override;
-    double Mode() const override;
+    int Mode() const override;
     double Skewness() const override;
     double ExcessKurtosis() const override;
