+++ distributions/univariate/discrete/ZipfRand.h
@@ -37,7 +37,7 @@ public:
     
     std::complex<double> CF(double t) const override;
      
-    double Mode() const override;
+    int Mode() const override;
     double Skewness() const override;
     double ExcessKurtosis() const override;
 };
